+++ packages/pangea-sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- IP Intel: modernized models.
+
 ## 6.5.0 - 2025-08-25
 
 ### Changed