+++ packages/pangea-sdk/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- AI Guard: improved `prompt_messages` typing.
 - IP Intel: modernized models.
 
 ## 6.5.0 - 2025-08-25