+++ packages/pangea-sdk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 6.8.0 - 2025-10-15
+
 ### Added
 
 - AI Guard: image detector override.